+++ crates/scheduler/src/call/crypto/ed25519.rs
@@ -31,7 +31,7 @@ use crate::Failure;
 use crate::applet::store::Memory;
 use crate::{DispatchSchedulerCall, SchedulerCall};
 
-pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
+pub(super) fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
         Api::IsSupported(call) => is_supported(call),
         Api::GetLayout(call) => or_fail!("board-api-crypto-ed25519", get_layout(call)),