+++ crates/scheduler/src/call/vendor.rs
@@ -25,7 +25,7 @@ use crate::SchedulerCall;
 #[cfg(feature = "board-api-vendor")]
 use crate::applet::store::StoreApi as _;
 
-pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
+pub(super) fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
         Api::Syscall(call) => or_fail!("board-api-vendor", syscall(call)),
     }