+++ crates/scheduler/src/event/timer.rs
@@ -19,7 +19,7 @@ use wasefire_error::Error;
 
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 #[derive_where(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Key<B: Board> {
+pub(crate) struct Key<B: Board> {
     pub timer: Id<board::Timer<B>>,
 }
 
@@ -36,10 +36,10 @@ impl<'a, B: Board> From<&'a Event<B>> for Key<B> {
 }
 
 impl<B: Board> Key<B> {
-    pub fn disable(self) -> Result<(), Error> {
+    pub(crate) fn disable(self) -> Result<(), Error> {
         use wasefire_board_api::timer::Api as _;
         board::Timer::<B>::disarm(self.timer)
     }
 }
 
-pub fn process() {}
+pub(crate) fn process() {}