+++ crates/scheduler/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Patch
 
+- Enable `unreachable_pub` lint
 - Disable `unused-features` lint
 - Use heterogeneous try blocks when needed
 - Update dependencies